refactor(likes): align file structure with bulletproof-react
