+++ src/features/likes/components/LikeButton/index.ts
